+++ odoo/local-src/specific_account/models/__init__.py
@@ -7,3 +7,4 @@ from . import account_invoice_report
 from . import res_company
 from . import res_config
 from . import sale_order
+from . import product_product